pyproject.toml
[tool.poetry]
name = "ucp-demo"
version = "0.1.0"
description = "Interactive demo showcasing the Universal Commerce Protocol (UCP)"
authors = ["UCP Demo Team"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.28.1"
websockets = "^15.0"
python-dotenv = "^1.0.0"
google-adk = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
